requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens
astunparse==1.6.3
async-lru==2.0.4
attrs==25.1.0
audioread==3.0.1
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blis==0.7.11
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpathlib==0.20.0
colorama
comm
confection==0.1.5
contourpy==1.3.1
coqpit-config==0.1.2
coqui-tts==0.25.3
coqui-tts-trainer==0.2.2
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
datasets==3.3.1
dateparser==1.1.8
debugpy
decorator
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docopt==0.6.2
dotenv==0.9.9
einops==0.8.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
encodec==0.1.1
exceptiongroup
executing
fastjsonschema==2.21.1
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
gast==0.6.0
gensim==4.3.3
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-texttospeech==2.25.0
google-pasta==0.2.0
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
importlib_metadata
inflect==7.5.0
ipykernel
ipython
ipywidgets==8.1.5
isoduration==20.11.0
jedi
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
Js2Py==0.74
json5==0.10.0
jsonlines==1.2.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client
jupyter_core
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.8.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.44.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline
mdurl==0.1.2
mistune==3.1.2
ml_dtypes==0.5.1
monotonic-alignment-search==0.1.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio
networkx==3.4.2
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
num2words==0.5.14
numba==0.61.0
numpy==1.26.4
openai==1.66.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
overrides==7.7.0
packaging
pandas==2.2.3
pandocfilters==1.5.1
parso
pickleshare
pillow==11.1.0
pipwin==0.5.2
platformdirs
pooch==1.8.2
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil
pure_eval
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments
pyjsparser==2.7.1
pyparsing==3.2.1
PyPrind==2.11.3
pysbd==0.3.4
pySmartDL==1.3.4
python-crfsuite==0.9.11
python-dateutil
python-dotenv==1.0.1
python-json-logger==3.3.0
pytz==2025.1
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.13.1
Send2Trash==1.8.3
setuptools==75.8.0
shellingham==1.5.4
six
smart-open==7.1.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack_data
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
tf_keras==2.19.0
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2025.3.13
tinycss2==1.4.0
tokenizers==0.20.3
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tornado
tqdm==4.67.1
traitlets
transformers==4.46.2
typeguard==4.4.2
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions
tzdata==2025.1
tzlocal==5.3
uri-template==1.3.0
urllib3==2.3.0
wasabi==1.1.3
wcwidth
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wget==3.2
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp
